--- a/python/benchmarks/version_chain.py
+++ b/python/benchmarks/version_chain.py
@@ -22,9 +22,8 @@ from .common import *
 
 
 class IterateVersionChain:
-    number = 10
+    number = 13
     timeout = 6000
-    warmup_time = 0
     rounds = 1
     CONNECTION_STRING = "lmdb://version_chain?map_size=20GB"
     LIB_NAME = "lib"