--- a/python/benchmarks/version_chain.py
+++ b/python/benchmarks/version_chain.py
@@ -36,7 +36,15 @@ class IterateVersionChain:
     def symbol(self, num_versions, deleted):
         return f"symbol_{num_versions}_{deleted}"
 
+    def __init__(self):
+        self.logger = get_logger()
+
     def setup_cache(self):
+        start = time.time()
+        self._setup_cache()
+        self.logger.info(f"SETUP_CACHE TIME: {time.time() - start}")
+
+    def _setup_cache(self):        
         self.ac = Arctic(IterateVersionChain.CONNECTION_STRING)
 
         num_versions_list, caching_list, deleted_list = IterateVersionChain.params