… python/benchmarks/persistent_query_builder.py