#### Reference Issues/PRs
<!--Example: Fixes #1234. See also #3456.-->

#### What does this implement or fix?

ASV test for finalize staged data on LMDB could be flaky. The reason was that it is executed exactly once as the staged data can be finalizes once.

With LMDB there is also conviniency that db is staored as files locally. Thus a solution where the database is created once, then backed up and then after each test it is copied back is implemented. This makes possible to have 5 executions which are more than sufficient to get stable avg, as finalization takes approx 8 secs (ie it is realtively slow and likelyhood to have problems vs situation where test is in ms is extremely low over 40 secs time, compared to situation where measurement times is < 1s

-------
Removed alse persistent_query_builder ASV tests it was source for flakiness (due to calculation based on 2 samples only) and it has been completely superseeded by: https://github.com/man-group/ArcticDB/blob/master/python/benchmarks/real_query_builder.py which has much more things to offer and utilizes fully the persistent storage.

NOTE: persistent_query_builder used to run with PRs, while the new one is overnight. This is the only downside but if needed can be included to run along LMDB tests

-------
version_chain - the only thing here that can be done with test is to pring back warm-up time and increase with 30% the number of samples gathered for the calculation

In fact I find this test extremely suitable for Real tests where we use Persistent storage for one time setup of data. As it costs 35 minutes only to setup LMDB this test is perfect candidate to be converted to such test. Moreover lets examine this log:
https://github.com/man-group/ArcticDB/actions/runs/17034135989/job/48282781563?pr=2590

```
| Change   | Before [99c0afa0]    | After [2d4962e0]    |   Ratio | Benchmark (Parameter)                                                                     |
|----------|----------------------|---------------------|---------|-------------------------------------------------------------------------------------------|
| +        | 98.2±4ms             | 142±20ms            |    1.45 | persistent_query_builder.PersistentQueryBuilderFunctions.time_query_adv_query_2(1000000)  |
| +        | 5.33±0.09ms          | 6.84±0.08ms         |    1.28 | version_chain.IterateVersionChain.time_read_alternating(25000, 'default', 0.0)            |
| +        | 5.43±0.08ms          | 6.51±0.3ms          |    1.2  | version_chain.IterateVersionChain.time_read_alternating(25000, 'forever', 0.0)            |
| +        | 5.42±0.09ms          | 6.41±0.08ms         |    1.18 | version_chain.IterateVersionChain.time_read_from_epoch(25000, 'default', 0.0)             |
| -        | 701±60ms             | 575±10ms            |    0.82 | persistent_query_builder.PersistentQueryBuilderFunctions.time_query_adv_query_2(10000000) |
```

It is seen that measurment for version_chain that are flaky are approx 6 ms each. Thus even with increased 30% number of samples from 10 to 13 we are going to experience further flakiness as both LMDB and Arcticdb are on the same machine and where times are in ms you can not rely that LMDB (or other related with it process) will not consume more proc time. Thus separating the storage and application under test is the right thing to do.


RESULTS FROM TESTS:

Run from ASV workflow without this PR: https://github.com/man-group/ArcticDB/actions/runs/17064710600/job/48379213172?pr=2594
2h 15m 29s

Run from this modified workflow: https://github.com/man-group/ArcticDB/actions/runs/17071349517/job/48401001814?pr=2590
2h 18m 11s

Just 3 minutes overall increase from 135 to 138 mins which is 2%

Additional run from Actions:
https://github.com/man-group/ArcticDB/actions/runs/17096787292

================================
Locally measured times for tests:

```
2025-08-21 11:01:16,531 - benchmarks.finalize_staged_data - INFO - SETUP_CACHE TIME: 25.863695859909058
Execution time: 81 seconds
--------------------
2025-08-21 11:03:59,690 - benchmarks.list_functions - INFO - SETUP_CACHE TIME: 58.31131935119629 
Execution time: 75 seconds
--------------------
2025-08-21 11:07:53,018 - benchmarks.list_snapshots - INFO - SETUP_CACHE TIME: 97.5234956741333
Execution time: 271 seconds
--------------------
2025-08-21 11:12:26,384 - benchmarks.local_query_builder - INFO - SETUP_CACHE TIME: 7.684977769851685
Execution time: 202 seconds
--------------------
2025-08-21 11:16:31,375 - benchmarks.bi_benchmarks - INFO - SETUP_CACHE TIME: 16.876145601272583
Execution time: 152 seconds
--------------------
2025-08-21 11:19:43,333 - benchmarks.basic_functions - INFO - SETUP_CACHE TIME: 9.34851861000061
BasicFunctions
Execution time: 284 seconds
--------------------
2025-08-21 11:31:22,862 - benchmarks.basic_functions - INFO - SETUP_CACHE TIME: 80.07152652740479
BatchBasicFunctions
Execution time: 1227 seconds
--------------------
2025-08-21 12:03:15,853 - benchmarks.basic_functions - INFO - SETUP_CACHE TIME: 11.19944167137146
ModificationFunctions
Execution time: 481 seconds
--------------------
2025-08-21 12:15:20,464 - benchmarks.arrow - INFO - SETUP_CACHE TIME: 17.815497636795044                                                                                                              arrow.ArrowReadNumeric
2025-08-21 13:19:27,823 - benchmarks.arrow - INFO - SETUP_CACHE TIME: 9.17694902420044
arrow.ArrowReadStrings
Execution time: 4075 seconds
```


#### Any other comments?

#### Checklist

<details>
  <summary>
   Checklist for code changes...
  </summary>
 
 - [ ] Have you updated the relevant docstrings, documentation and copyright notice?
 - [ ] Is this contribution tested against [all ArcticDB's features](../docs/mkdocs/docs/technical/contributing.md)?
 - [ ] Do all exceptions introduced raise appropriate [error messages](https://docs.arcticdb.io/error_messages/)?
 - [ ] Are API changes highlighted in the PR description?
 - [ ] Is the PR labelled as enhancement or bug so it appears in autogenerated release notes?
</details>

<!--
Thanks for contributing a Pull Request to ArcticDB! Please ensure you have taken a look at:
 - ArcticDB's Code of Conduct: https://github.com/man-group/ArcticDB/blob/master/CODE_OF_CONDUCT.md
 - ArcticDB's Contribution Licensing: https://github.com/man-group/ArcticDB/blob/master/docs/mkdocs/docs/technical/contributing.md#contribution-licensing
-->